services/feed-simulator/requirements.txt
redis>=5.0.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
